Cleanup and small rework of playback+recording code

+++ mod/recorder.py
@@ -1,4 +1,5 @@
 import time, subprocess, os, copy, json
+from signal import SIGINT
 from tornado import ioloop
 from mod.settings import CAPTURE_PATH, PLAYBACK_PATH
 
@@ -6,78 +7,48 @@ class Recorder(object):
     def __init__(self):
         self.recording = False
         self.tstamp = None
-        self.events = []
-        self.last_event = None
         self.proc = None
 
     def start(self):
         if self.recording:
-            self.stop()
+            self.stop(False)
         self.tstamp = time.time()
-        self.events = []
-        self.last_event = None
         self.proc = subprocess.Popen(['jack_capture',
                                       '-f', 'ogg',
-                                      '-V',
-                                      '-d', '65'],
-                                     stdout=open('/tmp/capture.err', 'w'),
-                                     stderr=open('/tmp/capture.out', 'w')
+                                      '-V', '-dc',
+                                      '-d', '65',
+                                      '--port', 'mod-host:monitor-out_1',
+                                      '--port', 'mod-host:monitor-out_2',
+                                      CAPTURE_PATH],
+                                     #stdout=open('/tmp/capture.err', 'w'),
+                                     #stderr=open('/tmp/capture.out', 'w')
                                      )
         self.recording = True
 
-    def stop(self):
+    def stop(self, returnFileHandle):
         if not self.recording:
-            return
-        self.proc.send_signal(2)
+            return None
+        self.proc.send_signal(SIGINT)
         self.proc.wait()
         self.recording = False
-        result = {
-            'handle': open(CAPTURE_PATH, 'rb'),
-            'events': copy.deepcopy(self.events),
-        }
+        fhandle = open(CAPTURE_PATH, 'rb') if returnFileHandle else None
         os.remove(CAPTURE_PATH)
-        self.events = []
-        self.last_event = None
-        return result
-
-    def event(self, event_type, *data):
-        if not self.recording:
-            return
-        fingerprint = json.dumps([event_type, data])
-        if self.last_event == fingerprint:
-            return
-        self.last_event = fingerprint
-        self.events.append({
-                'type': event_type,
-                'tstamp': time.time() - self.tstamp,
-                'data': data,
-                })
-
-    def bypass(self, instance_id, value):
-        self.event('bypass', instance_id, value)
-
-    def parameter(self, instance_id, port_id, value):
-        self.event('parameter', instance_id, port_id, value)
+        return fhandle
 
 class Player(object):
     def __init__(self):
         self.proc = None
-        self.fh = None
+        self.fhandle = None
         self.stop_callback = None
 
-    @property
-    def playing(self):
-        return self.proc is not None
-
-    def play(self, fh, stop_callback):
-        if self.playing:
-            self.stop()
-        fh.seek(0)
-        with open(PLAYBACK_PATH, 'wb') as fd:
-            fd.write(fh.read())
+    def play(self, fhandle, stop_callback):
+        self.stop()
+        fhandle.seek(0)
+        with open(PLAYBACK_PATH, 'wb') as fh:
+            fh.write(fhandle.read())
         self.proc = subprocess.Popen(['sndfile-jackplay', PLAYBACK_PATH],
                                      stdout=subprocess.PIPE)
-        self.fh = fh
+        self.fhandle = fhandle
         self.stop_callback = stop_callback
         ioloop.IOLoop().instance().add_handler(self.proc.stdout.fileno(), self.end_callback, 16)
 
@@ -86,19 +57,22 @@ class Player(object):
         if self.proc.poll() is None:
             return
         ioloop.IOLoop.instance().remove_handler(fileno)
-        self.fh.seek(0)
+        self.fhandle.seek(0)
         self.callback()
 
     def stop(self):
+        if self.proc is None:
+            return
         ioloop.IOLoop.instance().remove_handler(self.proc.stdout.fileno())
         if self.proc.poll() is None:
             self.proc.kill()
-        self.fh.seek(0)
         self.proc = None
+        self.fhandle.seek(0)
         self.callback()
 
     def callback(self):
+        if self.stop_callback is None:
+            return
         cb = self.stop_callback
-        if cb is not None:
-            self.stop_callback = None
-            cb()
+        self.stop_callback = None
+        cb()
